stabilize remote trigger plugin
Gets the remote job's build number and the remote job's location checking
the remote job's queue.

This implementation gets the response's header when the remote job is
triggered.

The response's header contains the queue's location of the remote job.
The queue's location contains the queue's number of the remote job.

After triggering the remote job and getting the response's header,
the queue of the remote job is supervised, failing the job if the remote job is
cancelled before being build or getting the remote job's location and
the remote job's build number if the remote job is build.

Change-Id: I65b6a95f653920d8185cc6bae0691f8e02017bba