Bug fix for remote job build number
A bug was reported because sometimes the "executable" flag in the
/queue/item/ query was not found, in that case, the remote job was
interpreted as cancelled.

The main problem was that the test scope did not cover all the use cases.

This bug fix, checks more attributes and flags in order to cover all the
use cases, and fails if the "executable" flag was not found.

Furthermore, "org.jenkins-ci.plugins" was upgraded, because the previous
version did not return the "executable" flag in the /queue/item/ query.

Change-Id: I6c74be14154c6e73c72b57af76579c3d89ce314b